requirements.txt
aiohttp
attrs
python-dateutil